+++ DentalOfficeManagement/src/main/java/com/utn/DentalOfficeManagement/Model/Pago.java
@@ -16,7 +16,7 @@ public class Pago {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_pago")
-    private Integer idPago;
+    private Long idPago;
 
     @Column(name = "saldo", nullable = false, precision = 10, scale = 2)
     private BigDecimal saldo;
